Update Remediate_Bitlocker.ps1
Changed the wait time until the key gets uploaded to Azure AD, before it waited 15 minutes (After the encrypt job is done) but now it does it almost directly after starting the encryption. I case someone closes laptop or shutdown device during encrypt phase the change is much lower now that the key is not in Azure AD yet.